requirements.txt
Django==2.2.1
django-rest-framework==0.1.0
django-rest-knox==4.0.1
djangorestframework==3.9.3
